Add CodecConfigObu processing logic to ProcessDescriptorObus.
This CL also adds a no-side-effect IsDataAvailable() function to read_bit_buffer.h. This helps us know when to stop reading data.

Along with the diffbase cl, this completes the proof-of-concept for the structure of processing a set of descriptor obus and sending them in a collection to the IamfDecoder.

Overall, the structure is like so (from top-level to lowest level):
1) IamfDecoder calls ProcessDescriptorObus
2) ProcessDescriptorObus calls header.ValidateAndRead to determine type of obu in question
3) ProcessDescriptorObus calls GetAndStore{ObuType} fxn for each obu of that type
4) GetAndStore{ObuType} calls {ObuType}::CreateFromBuffer, which returns a valid obu of that type.
5) {ObuType}::CreateFromBuffer calls {ObuType}::ValidateAndReadPayload, which reads the underlying payload data of the obu using our ReadBitBuffer
6) Once all descriptor obus have been processed and collected, ProcessDescriptorObus will call decoder.Set{ObuType} for each collection of obus.

PiperOrigin-RevId: 625360130