Add AudioFrameObu::CreateFromBuffer functionality.
This function will call a private explicit constructor which uses defaults to create an instance of the class. It will then call ValidateAndReadPayload, which will read from the provided buffer and fill in the appropriate fields within the AudioFrameObu instance. This call must succeed - if it does not, we will return an error status which will signal to the caller that we were not able to read the OBU properly.

This function will be called in ProcessTemporalUnitObus.

PiperOrigin-RevId: 631820105