Add a function to extract the relevant metadata from the UserMetadata text proto
associated with a given test vector.

For now, this is only used to properly associate gold-standard wav files with the output of the decoder. Currently, this means that we associate the first mix presentation with a stereo layout. In the future, we may also use this function to help determine what the inputs to the decoder should be in our tests (once we have the functionality to select mix presentations in the decoder) such that we produce the right wav files for all available gold-standard wav files.

PiperOrigin-RevId: 679699057