Add BeginTemporalUnit() to encoder API, removing side effect from GeneratingDataObus()
GeneratingDataObus was intended to be a const function
with no side effects, but that was not really true. This
change moves the side effect logic into its own function.

Note, this not a backwards compatible change, and any
user will need to add usage of BeginTemporalUnit()
in the right places.

PiperOrigin-RevId: 683671969