Add a class to house FLAC stream callback functions & implement the read callback.
These are necessary to use the FLAC stream decoder.

Also adds an encoded_frame_ field in flac_decoder.h which is used to store a single encoded frame at a time for later processing. This is necessary because flac stream uses callback functions. Getters & setters are added so that the call back functions can access the data as necessary.

PiperOrigin-RevId: 686183659