Implement FLAC decoder initialization + store the pointer to the stream decoder.
This CL also moves all the libflac callback functions to the FlacDecoder class; this is because in the current structure, there is a cyclic dependency between flac_decoder.h and flac_decoder_stream_callback.h. In order to break it, we add the callbacks back to flac_decoder.h. The tests for the callbacks are moved to the newly created flac_decoder_test.cc as well.

I'll remove flac_decoder_stream_callback.h & co in a separate cl.

PiperOrigin-RevId: 687382376