Implement DecodeAudioFrame in FlacDecoder.
This cl also fixes two bugs within LibFlacWriteCallback. The first is that planar libflac data (i.e. <channel<time_ticks>>) needs to be transposed to our interleaved representation (i.e. <time_tick<channels>>). The second is that we need to shift each sample left based on the bits_per_sample within a FLAC_int32, as it is not the case that their bits are always stored in the upper bits of a given int32. The tests for this fxn are updated accordingly.

PiperOrigin-RevId: 688601397