This CL begins the process of making the rendering portion of the decoder iterative.
Add structs to represent the rendering metadata that is necessary to keep track of our rendering logic. This cl also adds a new function to generate rendering metadata for submixes.

This function will be used to generate the rendering metadata for all submixes in a given mix presentation OBU. The rendering metadata will later be used to render the audio frames at each timestamp.

A subsequent cl will implement GenerateRenderingMetadataForAllLayouts().

PiperOrigin-RevId: 697736516