Implement Finalize() function for the iterative renderer.
With this function, the iterative renderer has the full functionality (namely, it can now be used to compute loudness) and can be used anywhere. Subsequent cls will update calls to the old renderer in the encoder with this new, iterative renderer. Once this is complete, we will remove and deprecate all the old code that currently exists in the renderer.

PiperOrigin-RevId: 702411796