Create `StreamBasedReadBitBuffer` class which extends the base `ReadBitBuffer` class.
This will be used by the iterative decoder in order to process data based on chunking provided by the caller. This cl provides the generic structure; subsequent cls will add implementation.

Sample usage pattern:
1) Create()
2) for chunk in stream:
   PushBytes(chunk)
   *ReadObu() & Flush() until not enough data available

Within read obu, we can use tell() at the beginning of the function (after reading the obu header) to see if there is enough data to read a whole obu, since we already know the expected obu size at this point.

If tell signals that we do not have enough data, we use seek to reset the buffer to before having read the header & also ask the caller to provide more data via PushBytes().

If tell signals we do have enough data, then we simply read and process the obu accordingly.

*Note that ReadObu() is not a function within the read bit buffer. This refers to an obu processing function (of which there are many) which calls some combination of reading methods of the read bit buffer class.

PiperOrigin-RevId: 707614225