Add helper for peeking at the OBU type and total OBU size.
This function is necessary for obu processing of binary streams. It will be used in the processing of descriptor obus to check for the existence of a temporal unit header.

This function is a strict Peek() function, meaning that no matter whether it succeeds or fails, it will never consume data from the input buffer. It also establishes a contract for failure due to a lack of data - in this case, it will always return an absl::ResourceExhausted() error. Client functions will rely on this contract.

PiperOrigin-RevId: 720987694