Remove dependency on wav writing in obu processor.
This is necessary so the iterative decoder can use this class appropriately, since in the iterative case, we do not do any wav writing, as samples are processed and delivered to the client as is.

This cl updates the obu processor to take in a SampleProcessorFactory instead of creating a WavWriter, which allows for the removal of the wav writer dependency from this module.

The behavior of the processor remains unchanged. Two tests that check the overriding of bit depths for a wav file are removed, since they are no longer in scope of this module - this feature is tested in the rendering mix presentation finalizer.

PiperOrigin-RevId: 725259810