Surface outputted pcm samples from the renderer to the top-level of the obu processor.
In order to do this, we also store the target layout information when we create the renderer as well; this information is then used to call GetPostProcessedSamplesAsSpan.

In the case that wav writers are used, the new output argument has no effect and will be empty; we update the call in api/google/iamf_decoder to pass in this unused arg.

PiperOrigin-RevId: 725667048