Implement GetOutputTemporalUnit().
This function returns the first temporal unit that was decoded. Because of this, we update the data structure which holds the rendered pcm samples to be a queue on the outermost layer, where each element represents a decoded temporal unit. This gives us O(1) time complexity on the querying of temporal units.

PiperOrigin-RevId: 728772932