Implement IamfDecoder::Flush.
In order to support this, we add a new state_ enum which tracks the state of the decoder. It ensures that Decode cannot be called after a call to flush has been made. Flush is designed to be continually called until it signals it has output all remaining temporal units.

PiperOrigin-RevId: 729162659