Implement `IamfDecoder::GetOutputLayout()`.
This allows the user to query which output layout will be used to render the audio, as it might be different than what they requested if the desired layout was not found in the mix presentations of the descriptors. This can only be called after the descriptors have been processed.

PiperOrigin-RevId: 739237169