Add the ability to Reset the IamfDecoder.
The effect of this function is to reset the decoder to a post-descriptors processed, pre-temporal unit state in which decoding can begin anew with the same descriptors. Naturally, it is therefore not allowed to call this function without descriptors having already been processed, i.e. IsDescriptorProcessingComplete() must return true.

This cl also renames SignalEndOfStream to SignalEndOfDecoding, as this may be called before the end of the stream.

Finally, this cl removes an unused input bitstream arg in CreateObuProcessor and adds an output_descriptors arg which is populated with the serialized descriptors when they are being originally processed - we then use these to recreate a fresh obu processor when reset is called to ensure all the stateful information in the obu processor is aligned.

PiperOrigin-RevId: 743219819