Add MixPresentation ID as a means to select the output
This change refactors the IAMF decoder API to allow clients to select a specific mix presentation, by ID, from the bitstream, in addition to or instead of the desired output layout.

This is a breaking change that includes the following:

The `requested_layout` field in `IamfDecoderFactory::Settings` is replaced by a `requested_mix` field of type `RequestedMix`.

The `GetOutputLayout` function is replaced with `GetOutputMix`, which returns the selected mix presentation ID and the resulting output layout in a `SelectedMix` struct.

The `ResetWithNewLayout` function is replaced with `ResetWithNewMix`.

PiperOrigin-RevId: 794674387