Add DualPolarParamDefinition.
This CL introduces the DualPolarParamDefinition class, which is a subclass of ParamDefinition. It is used to define parameters for dual polar information.

The class includes methods for validating and writing to a bit buffer, reading from a bit buffer, creating a DualPolarParameterData instance, and printing the definition. It stores default values for first and second azimuth, elevation, and distance.

PiperOrigin-RevId: 843327051