Add DualCart16ParamDefinition class.
This CL introduces the `DualCart16ParamDefinition` class, which is a subclass of `ParamDefinition`. It is used to define parameters of type `kParameterDefinitionDualCart16`. The class includes default values for two sets of 3D Cartesian coordinates (x, y, z), each stored as `int16_t`. It provides methods to validate and write, read and validate, create associated parameter data, and print the definition.

PiperOrigin-RevId: 844486539